requirements.txt
accelerate==1.0.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
av==14.0.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
datasets==2.19.1
deepspeed==0.16.2
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
et_xmlfile==2.0.0
evaluate==0.4.1
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.13.1
fire==0.7.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.44
gradio==4.44.1
gradio_client==1.3.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.27.1
idna==3.10
imageio==2.36.1
importlib_resources==6.5.2
jieba==0.42.1
Jinja2==3.1.3
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
monai==1.3.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
nibabel==5.2.1
ninja==1.11.1.3
nltk==3.9.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
ollama==0.4.5
openai==1.59.6
opencv-python==4.9.0.80
openpyxl==3.1.5
orjson==3.10.13
pandas==2.2.3
peft==0.12.0
pillow==10.3.0
propcache==0.2.1
protobuf==5.29.2
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyarrow-hotfix==0.6
pycocotools==2.0.7
pydantic==2.10.4
pydantic_core==2.27.2
pydicom==3.0.1
pydub==0.25.1
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
pyqtgraph==0.13.7
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
qimage2ndarray==1.10.0
QtAwesome==1.3.1
QtPy==2.4.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.9.4
rouge==1.0.1
rouge-chinese==1.0.3
ruff==0.8.6
safetensors==0.4.3
scipy==1.15.0
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
shtab==1.7.1
simple-slice-viewer==0.97
SimpleITK==2.3.1
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.41.3
sympy==1.13.1
tensorboardX==2.6.2.2
termcolor==2.5.0
tiktoken==0.8.0
tokenizers==0.20.3
tomlkit==0.12.0
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
tqdm==4.66.2
transformers==4.46.1
triton==2.2.0
trl==0.9.6
tweepy==4.14.0
typer==0.15.1
tyro==0.8.14
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
wandb==0.19.2
websockets==12.0
xxhash==3.5.0
yarl==1.18.3
